app/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
asyncpg==0.29.0
redis==5.0.1
httpx==0.25.2
celery==5.3.4
structlog==23.2.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
huggingface-hub==0.19.4
qdrant-client==1.6.0
watchdog==3.0.0
APScheduler==3.10.4
python-dateutil==2.8.2
prometheus-client==0.17.1
pytz==2023.3.post1
PyYAML==6.0.1
pywebpush==1.14.0
python-multipart==0.0.6
